move settings to the side-side-panel
